Create pivot list

Create pivot list

Usage

create_pivot_list(pivot_data, arrange_cols = "process")

Arguments

pivot_data

privot data as retrieved from function pivot_data()

arrange_cols

columns used for arranging the data (default: "process")

Value

a list of results, where each element contains the result table for one lci_method

Examples

zipfile <- system.file("extdata/umberto-nxt_v7.1.0.13.503/Beispiel_Auswertung.zip", 
package = "kwb.umberto")
umberto7_csv_dir <- file.path(tempdir(), "Beispiel_Auswertung")
unzip(zipfile, exdir = umberto7_csv_dir)
umberto7_rawdata <- kwb.umberto::import_rawdata(csv_dir = umberto7_csv_dir)
#> Importing csv file 'C:\Users\RUNNER~1\AppData\Local\Temp\RtmpMNHWtj/Beispiel_Auswertung/Ref_High_C+N.csv'
#> ℹ Using "','" as decimal and "'.'" as grouping mark. Use `read_delim()` for more control.
#> Rows: 19996 Columns: 12
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ";"
#> chr (11): Project, Model, Net, Timestamp, Product, LCI Method, Phase, Proces...
#> dbl  (1): Quantity
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Importing csv file 'C:\Users\RUNNER~1\AppData\Local\Temp\RtmpMNHWtj/Beispiel_Auswertung/Ref_High_C+high N.csv'
#> ℹ Using "','" as decimal and "'.'" as grouping mark. Use `read_delim()` for more control.
#> Rows: 19996 Columns: 12
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ";"
#> chr (11): Project, Model, Net, Timestamp, Product, LCI Method, Phase, Proces...
#> dbl  (1): Quantity
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Importing csv file 'C:\Users\RUNNER~1\AppData\Local\Temp\RtmpMNHWtj/Beispiel_Auswertung/Ref_Low_C+N.csv'
#> ℹ Using "','" as decimal and "'.'" as grouping mark. Use `read_delim()` for more control.
#> Rows: 19996 Columns: 12
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ";"
#> chr (11): Project, Model, Net, Timestamp, Product, LCI Method, Phase, Proces...
#> dbl  (1): Quantity
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Importing csv file 'C:\Users\RUNNER~1\AppData\Local\Temp\RtmpMNHWtj/Beispiel_Auswertung/Ref_Low_C+high N.csv'
#> ℹ Using "','" as decimal and "'.'" as grouping mark. Use `read_delim()` for more control.
#> Rows: 19996 Columns: 12
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ";"
#> chr (11): Project, Model, Net, Timestamp, Product, LCI Method, Phase, Proces...
#> dbl  (1): Quantity
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
umberto7_data_grouped <- kwb.umberto::group_data(umberto7_rawdata)
umberto7_data_pivot <- kwb.umberto::pivot_data(umberto7_data_grouped)
umberto7_data_pivot_list <- kwb.umberto::create_pivot_list(umberto7_data_pivot)
#> Warning: `arrange_()` was deprecated in dplyr 0.7.0.
#> ℹ Please use `arrange()` instead.
#> ℹ See vignette('programming') for more help
#> ℹ The deprecated feature was likely used in the dplyr package.
#>   Please report the issue at <https://github.com/tidyverse/dplyr/issues>.
#> Joining, by = c("lci_method", "process")
#> Joining, by = c("lci_method", "process")
#> Joining, by = c("lci_method", "process")
#> Joining, by = c("lci_method", "process")
#> Joining, by = c("lci_method", "process")
#> Joining, by = c("lci_method", "process")
#> Joining, by = c("lci_method", "process")
#> Joining, by = c("lci_method", "process")
#> Joining, by = c("lci_method", "process")
head(umberto7_data_pivot_list)
#> $lci_method1
#> # A tibble: 14 × 6
#> # Groups:   lci_method, process [14]
#>    lci_method                            process 0a_Re…¹ 0b_Re…² 0c_Re…³ 0d_Re…⁴
#>    <chr>                                 <chr>     <dbl>   <dbl>   <dbl>   <dbl>
#>  1 ReCiPe Midpoint (H) w/o LT - climate… T03: A…  25498.   6656.  17548.   2705.
#>  2 ReCiPe Midpoint (H) w/o LT - climate… T06: C…    607.    604.    615.    608.
#>  3 ReCiPe Midpoint (H) w/o LT - climate… T07: C…    277.    281.    284.    282.
#>  4 ReCiPe Midpoint (H) w/o LT - climate… T10: E…     NA      NA      NA      NA 
#>  5 ReCiPe Midpoint (H) w/o LT - climate… T11: E…  12520.  11694.  13973.  13312.
#>  6 ReCiPe Midpoint (H) w/o LT - climate… T12: E…   2602.   2556.   2666.   2569.
#>  7 ReCiPe Midpoint (H) w/o LT - climate… T13: F…   1548.   2016.   2264.   2085.
#>  8 ReCiPe Midpoint (H) w/o LT - climate… T14: P…    807.    834.    823.    830.
#>  9 ReCiPe Midpoint (H) w/o LT - climate… T15: S…   1382.   1423.   1443.   1440.
#> 10 ReCiPe Midpoint (H) w/o LT - climate… T21: M…   6154.   6308.   6391.   6372.
#> 11 ReCiPe Midpoint (H) w/o LT - climate… T30: H…  -1262.  -1207.  -1197.  -1184.
#> 12 ReCiPe Midpoint (H) w/o LT - climate… T31: E… -12697. -12457. -12596. -12443.
#> 13 ReCiPe Midpoint (H) w/o LT - climate… T39: t…   1683.   1822.   1875.   1880.
#> 14 ReCiPe Midpoint (H) w/o LT - climate… T51: I…   2134.   1778.   2199.   1585.
#> # … with abbreviated variable names ¹​`0a_Ref_Low_C+N`, ²​`0b_Ref_High_C+N`,
#> #   ³​`0c_Ref_Low_C+high N`, ⁴​`0d_Ref_High_C+high N`
#> 
#> $lci_method2
#> # A tibble: 14 × 6
#> # Groups:   lci_method, process [14]
#>    lci_method                            process 0a_Re…¹ 0b_Re…² 0c_Re…³ 0d_Re…⁴
#>    <chr>                                 <chr>     <dbl>   <dbl>   <dbl>   <dbl>
#>  1 ReCiPe Midpoint (H) w/o LT - freshwa… T03: A…  NA      NA      NA      NA    
#>  2 ReCiPe Midpoint (H) w/o LT - freshwa… T06: C…  NA      NA      NA      NA    
#>  3 ReCiPe Midpoint (H) w/o LT - freshwa… T07: C…  NA      NA      NA      NA    
#>  4 ReCiPe Midpoint (H) w/o LT - freshwa… T10: E…  NA      NA      NA      NA    
#>  5 ReCiPe Midpoint (H) w/o LT - freshwa… T11: E…   3.24    3.03    3.62    3.45 
#>  6 ReCiPe Midpoint (H) w/o LT - freshwa… T12: E…   0.673   0.662   0.690   0.665
#>  7 ReCiPe Midpoint (H) w/o LT - freshwa… T13: F…   1.27    1.65    1.85    1.70 
#>  8 ReCiPe Midpoint (H) w/o LT - freshwa… T14: P…   0.558   0.577   0.569   0.575
#>  9 ReCiPe Midpoint (H) w/o LT - freshwa… T15: S…   1.07    1.10    1.12    1.12 
#> 10 ReCiPe Midpoint (H) w/o LT - freshwa… T21: M…   0.207   0.212   0.215   0.214
#> 11 ReCiPe Midpoint (H) w/o LT - freshwa… T30: H…  -0.526  -0.503  -0.499  -0.493
#> 12 ReCiPe Midpoint (H) w/o LT - freshwa… T31: E…  -3.29   -3.22   -3.26   -3.22 
#> 13 ReCiPe Midpoint (H) w/o LT - freshwa… T39: t…   0.883   0.956   0.984   0.986
#> 14 ReCiPe Midpoint (H) w/o LT - freshwa… T51: I…   0.929   0.774   0.958   0.690
#> # … with abbreviated variable names ¹​`0a_Ref_Low_C+N`, ²​`0b_Ref_High_C+N`,
#> #   ³​`0c_Ref_Low_C+high N`, ⁴​`0d_Ref_High_C+high N`
#> 
#> $lci_method3
#> # A tibble: 14 × 6
#> # Groups:   lci_method, process [14]
#>    lci_method                        process 0a_Ref…¹ 0b_Ref…² 0c_Ref…³ 0d_Ref…⁴
#>    <chr>                             <chr>      <dbl>    <dbl>    <dbl>    <dbl>
#>  1 ReCiPe Midpoint (H) w/o LT - fre… T03: A… NA       NA       NA       NA      
#>  2 ReCiPe Midpoint (H) w/o LT - fre… T06: C… NA       NA       NA       NA      
#>  3 ReCiPe Midpoint (H) w/o LT - fre… T07: C… NA       NA       NA       NA      
#>  4 ReCiPe Midpoint (H) w/o LT - fre… T10: E…  1.06e+2 53.2     32.3     16.2    
#>  5 ReCiPe Midpoint (H) w/o LT - fre… T11: E…  1.21e+0  1.13     1.35     1.28   
#>  6 ReCiPe Midpoint (H) w/o LT - fre… T12: E…  2.51e-1  0.246    0.257    0.247  
#>  7 ReCiPe Midpoint (H) w/o LT - fre… T13: F…  2.21e-1  0.288    0.324    0.298  
#>  8 ReCiPe Midpoint (H) w/o LT - fre… T14: P…  1.89e-2  0.0196   0.0193   0.0195 
#>  9 ReCiPe Midpoint (H) w/o LT - fre… T15: S…  1.99e-2  0.0205   0.0208   0.0207 
#> 10 ReCiPe Midpoint (H) w/o LT - fre… T21: M…  1.56e-2  0.0160   0.0162   0.0162 
#> 11 ReCiPe Midpoint (H) w/o LT - fre… T30: H… -9.95e-3 -0.00951 -0.00943 -0.00933
#> 12 ReCiPe Midpoint (H) w/o LT - fre… T31: E… -1.22e+0 -1.20    -1.21    -1.20   
#> 13 ReCiPe Midpoint (H) w/o LT - fre… T39: t…  5.63e-2  0.0609   0.0627   0.0629 
#> 14 ReCiPe Midpoint (H) w/o LT - fre… T51: I…  1.12e-1  0.0936   0.116    0.0834 
#> # … with abbreviated variable names ¹​`0a_Ref_Low_C+N`, ²​`0b_Ref_High_C+N`,
#> #   ³​`0c_Ref_Low_C+high N`, ⁴​`0d_Ref_High_C+high N`
#> 
#> $lci_method4
#> # A tibble: 14 × 6
#> # Groups:   lci_method, process [14]
#>    lci_method                            process 0a_Re…¹ 0b_Re…² 0c_Re…³ 0d_Re…⁴
#>    <chr>                                 <chr>     <dbl>   <dbl>   <dbl>   <dbl>
#>  1 ReCiPe Midpoint (H) w/o LT - human t… T03: A…    NA      NA      NA      NA  
#>  2 ReCiPe Midpoint (H) w/o LT - human t… T06: C…    NA      NA      NA      NA  
#>  3 ReCiPe Midpoint (H) w/o LT - human t… T07: C…    NA      NA      NA      NA  
#>  4 ReCiPe Midpoint (H) w/o LT - human t… T10: E…    NA      NA      NA      NA  
#>  5 ReCiPe Midpoint (H) w/o LT - human t… T11: E…   853.    797.    952.    907. 
#>  6 ReCiPe Midpoint (H) w/o LT - human t… T12: E…   177.    174.    182.    175. 
#>  7 ReCiPe Midpoint (H) w/o LT - human t… T13: F…   355.    463.    520.    479. 
#>  8 ReCiPe Midpoint (H) w/o LT - human t… T14: P…    52.2    54.0    53.2    53.7
#>  9 ReCiPe Midpoint (H) w/o LT - human t… T15: S…   292.    300.    305.    304. 
#> 10 ReCiPe Midpoint (H) w/o LT - human t… T21: M…    38.4    39.4    39.9    39.8
#> 11 ReCiPe Midpoint (H) w/o LT - human t… T30: H…   -18.4   -17.6   -17.4   -17.2
#> 12 ReCiPe Midpoint (H) w/o LT - human t… T31: E…  -865.   -849.   -858.   -848. 
#> 13 ReCiPe Midpoint (H) w/o LT - human t… T39: t…   141.    153.    157.    157. 
#> 14 ReCiPe Midpoint (H) w/o LT - human t… T51: I…   265.    221.    273.    197. 
#> # … with abbreviated variable names ¹​`0a_Ref_Low_C+N`, ²​`0b_Ref_High_C+N`,
#> #   ³​`0c_Ref_Low_C+high N`, ⁴​`0d_Ref_High_C+high N`
#> 
#> $lci_method5
#> # A tibble: 14 × 6
#> # Groups:   lci_method, process [14]
#>    lci_method                            process 0a_Re…¹ 0b_Re…² 0c_Re…³ 0d_Re…⁴
#>    <chr>                                 <chr>     <dbl>   <dbl>   <dbl>   <dbl>
#>  1 ReCiPe Midpoint (H) w/o LT - marine … T03: A…  NA      NA      NA      NA    
#>  2 ReCiPe Midpoint (H) w/o LT - marine … T06: C…  NA      NA      NA      NA    
#>  3 ReCiPe Midpoint (H) w/o LT - marine … T07: C…  NA      NA      NA      NA    
#>  4 ReCiPe Midpoint (H) w/o LT - marine … T10: E…  NA      NA      NA      NA    
#>  5 ReCiPe Midpoint (H) w/o LT - marine … T11: E…   5.33    4.98    5.95    5.67 
#>  6 ReCiPe Midpoint (H) w/o LT - marine … T12: E…   1.11    1.09    1.14    1.09 
#>  7 ReCiPe Midpoint (H) w/o LT - marine … T13: F…   3.63    4.72    5.31    4.89 
#>  8 ReCiPe Midpoint (H) w/o LT - marine … T14: P…   0.801   0.829   0.817   0.825
#>  9 ReCiPe Midpoint (H) w/o LT - marine … T15: S…   4.91    5.05    5.13    5.11 
#> 10 ReCiPe Midpoint (H) w/o LT - marine … T21: M…   0.508   0.520   0.527   0.526
#> 11 ReCiPe Midpoint (H) w/o LT - marine … T30: H…  -0.569  -0.544  -0.539  -0.534
#> 12 ReCiPe Midpoint (H) w/o LT - marine … T31: E…  -5.41   -5.30   -5.36   -5.30 
#> 13 ReCiPe Midpoint (H) w/o LT - marine … T39: t…   1.34    1.45    1.49    1.50 
#> 14 ReCiPe Midpoint (H) w/o LT - marine … T51: I…   2.22    1.85    2.28    1.65 
#> # … with abbreviated variable names ¹​`0a_Ref_Low_C+N`, ²​`0b_Ref_High_C+N`,
#> #   ³​`0c_Ref_Low_C+high N`, ⁴​`0d_Ref_High_C+high N`
#> 
#> $lci_method6
#> # A tibble: 14 × 6
#> # Groups:   lci_method, process [14]
#>    lci_method                        process 0a_Ref…¹ 0b_Ref…² 0c_Ref…³ 0d_Ref…⁴
#>    <chr>                             <chr>      <dbl>    <dbl>    <dbl>    <dbl>
#>  1 ReCiPe Midpoint (H) w/o LT - mar… T03: A…  3.79e+0   3.78    3.81e+0   3.78  
#>  2 ReCiPe Midpoint (H) w/o LT - mar… T06: C…  3.25e-1   0.324   3.29e-1   0.325 
#>  3 ReCiPe Midpoint (H) w/o LT - mar… T07: C… NA        NA      NA        NA     
#>  4 ReCiPe Midpoint (H) w/o LT - mar… T10: E…  1.46e+3 730.      1.10e+3 551.    
#>  5 ReCiPe Midpoint (H) w/o LT - mar… T11: E…  1.03e+0   0.964   1.15e+0   1.10  
#>  6 ReCiPe Midpoint (H) w/o LT - mar… T12: E…  2.14e-1   0.211   2.20e-1   0.212 
#>  7 ReCiPe Midpoint (H) w/o LT - mar… T13: F…  2.51e-1   0.326   3.67e-1   0.338 
#>  8 ReCiPe Midpoint (H) w/o LT - mar… T14: P…  2.97e+0   3.07    3.03e+0   3.06  
#>  9 ReCiPe Midpoint (H) w/o LT - mar… T15: S…  1.74e-1   0.179   1.82e-1   0.181 
#> 10 ReCiPe Midpoint (H) w/o LT - mar… T21: M…  2.50e-1   0.256   2.60e-1   0.259 
#> 11 ReCiPe Midpoint (H) w/o LT - mar… T30: H… -4.32e-2  -0.0413 -4.10e-2  -0.0405
#> 12 ReCiPe Midpoint (H) w/o LT - mar… T31: E… -1.05e+0  -1.03   -1.04e+0  -1.03  
#> 13 ReCiPe Midpoint (H) w/o LT - mar… T39: t…  1.80e-1   0.195   2.01e-1   0.201 
#> 14 ReCiPe Midpoint (H) w/o LT - mar… T51: I…  2.06e-1   0.171   2.12e-1   0.153 
#> # … with abbreviated variable names ¹​`0a_Ref_Low_C+N`, ²​`0b_Ref_High_C+N`,
#> #   ³​`0c_Ref_Low_C+high N`, ⁴​`0d_Ref_High_C+high N`
#> 

umberto10_csv_dir <- system.file("extdata/umberto-lca+_v10.1.0.3.146", 
package = "kwb.umberto")
umberto10_rawdata <- kwb.umberto::import_rawdata(csv_dir = umberto10_csv_dir)
#> Importing csv file 'D:/a/_temp/Library/kwb.umberto/extdata/umberto-lca+_v10.1.0.3.146/smartech2_model-v0.1.0_input-v0.3.1.csv'
#> ℹ Using "','" as decimal and "'.'" as grouping mark. Use `read_delim()` for more control.
#> Rows: 8456 Columns: 15
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ";"
#> chr (14): Project, Model, Net, Timestamp, Product, Product Name, Product Arr...
#> dbl  (1): Quantity
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
umberto10_data_grouped <- kwb.umberto::group_data(umberto10_rawdata)
umberto10_data_pivot <- kwb.umberto::pivot_data(umberto10_data_grouped)
umberto10_data_pivot_list <- kwb.umberto::create_pivot_list(umberto10_data_pivot)
#> Joining, by = c("lci_method", "process")
#> Joining, by = c("lci_method", "process")
#> Joining, by = c("lci_method", "process")
#> Joining, by = c("lci_method", "process")
#> Joining, by = c("lci_method", "process")
#> Joining, by = c("lci_method", "process")
#> Joining, by = c("lci_method", "process")
#> Joining, by = c("lci_method", "process")
#> Joining, by = c("lci_method", "process")
head(umberto10_data_pivot_list)
#> $lci_method1
#> # A tibble: 9 × 3
#> # Groups:   lci_method, process [9]
#>   lci_method                                                     process 0_Ref…¹
#>   <chr>                                                          <chr>     <dbl>
#> 1 ReCiPe Midpoint (H) w/o LT - climate change w/o LT, GWP100 w/… T03: A…  3.78e6
#> 2 ReCiPe Midpoint (H) w/o LT - climate change w/o LT, GWP100 w/… T06: C…  1.81e5
#> 3 ReCiPe Midpoint (H) w/o LT - climate change w/o LT, GWP100 w/… T07: C…  2.89e5
#> 4 ReCiPe Midpoint (H) w/o LT - climate change w/o LT, GWP100 w/… T10: E… NA     
#> 5 ReCiPe Midpoint (H) w/o LT - climate change w/o LT, GWP100 w/… T14: P…  3.93e5
#> 6 ReCiPe Midpoint (H) w/o LT - climate change w/o LT, GWP100 w/… T15: S…  1.10e2
#> 7 ReCiPe Midpoint (H) w/o LT - climate change w/o LT, GWP100 w/… T21: f… -2.17e5
#> 8 ReCiPe Midpoint (H) w/o LT - climate change w/o LT, GWP100 w/… T31: E… -1.17e6
#> 9 ReCiPe Midpoint (H) w/o LT - climate change w/o LT, GWP100 w/… T4: El…  1.33e7
#> # … with abbreviated variable name ¹​`0_Reference_Agri`
#> 
#> $lci_method2
#> # A tibble: 9 × 3
#> # Groups:   lci_method, process [9]
#>   lci_method                                                    process 0_Refe…¹
#>   <chr>                                                         <chr>      <dbl>
#> 1 ReCiPe Midpoint (H) w/o LT - freshwater ecotoxicity w/o LT, … T03: A… NA      
#> 2 ReCiPe Midpoint (H) w/o LT - freshwater ecotoxicity w/o LT, … T06: C… NA      
#> 3 ReCiPe Midpoint (H) w/o LT - freshwater ecotoxicity w/o LT, … T07: C… NA      
#> 4 ReCiPe Midpoint (H) w/o LT - freshwater ecotoxicity w/o LT, … T10: E… NA      
#> 5 ReCiPe Midpoint (H) w/o LT - freshwater ecotoxicity w/o LT, … T14: P…  2.81e+2
#> 6 ReCiPe Midpoint (H) w/o LT - freshwater ecotoxicity w/o LT, … T15: S…  8.43e-2
#> 7 ReCiPe Midpoint (H) w/o LT - freshwater ecotoxicity w/o LT, … T21: f… -2.76e+3
#> 8 ReCiPe Midpoint (H) w/o LT - freshwater ecotoxicity w/o LT, … T31: E… -3.08e+2
#> 9 ReCiPe Midpoint (H) w/o LT - freshwater ecotoxicity w/o LT, … T4: El…  2.06e+3
#> # … with abbreviated variable name ¹​`0_Reference_Agri`
#> 
#> $lci_method3
#> # A tibble: 9 × 3
#> # Groups:   lci_method, process [9]
#>   lci_method                                                    process 0_Refe…¹
#>   <chr>                                                         <chr>      <dbl>
#> 1 ReCiPe Midpoint (H) w/o LT - freshwater eutrophication w/o L… T03: A… NA      
#> 2 ReCiPe Midpoint (H) w/o LT - freshwater eutrophication w/o L… T06: C… NA      
#> 3 ReCiPe Midpoint (H) w/o LT - freshwater eutrophication w/o L… T07: C… NA      
#> 4 ReCiPe Midpoint (H) w/o LT - freshwater eutrophication w/o L… T10: E…  3.01e+3
#> 5 ReCiPe Midpoint (H) w/o LT - freshwater eutrophication w/o L… T14: P…  8.99e+0
#> 6 ReCiPe Midpoint (H) w/o LT - freshwater eutrophication w/o L… T15: S…  1.60e-3
#> 7 ReCiPe Midpoint (H) w/o LT - freshwater eutrophication w/o L… T21: f…  1.86e+3
#> 8 ReCiPe Midpoint (H) w/o LT - freshwater eutrophication w/o L… T31: E… -5.41e+1
#> 9 ReCiPe Midpoint (H) w/o LT - freshwater eutrophication w/o L… T4: El…  7.70e+2
#> # … with abbreviated variable name ¹​`0_Reference_Agri`
#> 
#> $lci_method4
#> # A tibble: 9 × 3
#> # Groups:   lci_method, process [9]
#>   lci_method                                                     process 0_Ref…¹
#>   <chr>                                                          <chr>     <dbl>
#> 1 ReCiPe Midpoint (H) w/o LT - human toxicity w/o LT, HTPinf w/… T03: A… NA     
#> 2 ReCiPe Midpoint (H) w/o LT - human toxicity w/o LT, HTPinf w/… T06: C… NA     
#> 3 ReCiPe Midpoint (H) w/o LT - human toxicity w/o LT, HTPinf w/… T07: C… NA     
#> 4 ReCiPe Midpoint (H) w/o LT - human toxicity w/o LT, HTPinf w/… T10: E… NA     
#> 5 ReCiPe Midpoint (H) w/o LT - human toxicity w/o LT, HTPinf w/… T14: P…  2.38e4
#> 6 ReCiPe Midpoint (H) w/o LT - human toxicity w/o LT, HTPinf w/… T15: S…  2.32e1
#> 7 ReCiPe Midpoint (H) w/o LT - human toxicity w/o LT, HTPinf w/… T21: f… -1.10e5
#> 8 ReCiPe Midpoint (H) w/o LT - human toxicity w/o LT, HTPinf w/… T31: E… -7.52e4
#> 9 ReCiPe Midpoint (H) w/o LT - human toxicity w/o LT, HTPinf w/… T4: El…  5.72e5
#> # … with abbreviated variable name ¹​`0_Reference_Agri`
#> 
#> $lci_method5
#> # A tibble: 9 × 3
#> # Groups:   lci_method, process [9]
#>   lci_method                                                    process 0_Refe…¹
#>   <chr>                                                         <chr>      <dbl>
#> 1 ReCiPe Midpoint (H) w/o LT - marine ecotoxicity w/o LT, METP… T03: A… NA      
#> 2 ReCiPe Midpoint (H) w/o LT - marine ecotoxicity w/o LT, METP… T06: C… NA      
#> 3 ReCiPe Midpoint (H) w/o LT - marine ecotoxicity w/o LT, METP… T07: C… NA      
#> 4 ReCiPe Midpoint (H) w/o LT - marine ecotoxicity w/o LT, METP… T10: E… NA      
#> 5 ReCiPe Midpoint (H) w/o LT - marine ecotoxicity w/o LT, METP… T14: P…  3.85e+2
#> 6 ReCiPe Midpoint (H) w/o LT - marine ecotoxicity w/o LT, METP… T15: S…  3.91e-1
#> 7 ReCiPe Midpoint (H) w/o LT - marine ecotoxicity w/o LT, METP… T21: f… -1.82e+3
#> 8 ReCiPe Midpoint (H) w/o LT - marine ecotoxicity w/o LT, METP… T31: E… -5.49e+2
#> 9 ReCiPe Midpoint (H) w/o LT - marine ecotoxicity w/o LT, METP… T4: El…  3.17e+3
#> # … with abbreviated variable name ¹​`0_Reference_Agri`
#> 
#> $lci_method6
#> # A tibble: 9 × 3
#> # Groups:   lci_method, process [9]
#>   lci_method                                                    process 0_Refe…¹
#>   <chr>                                                         <chr>      <dbl>
#> 1 ReCiPe Midpoint (H) w/o LT - marine eutrophication w/o LT, M… T03: A…  5.92e+2
#> 2 ReCiPe Midpoint (H) w/o LT - marine eutrophication w/o LT, M… T06: C…  9.67e+1
#> 3 ReCiPe Midpoint (H) w/o LT - marine eutrophication w/o LT, M… T07: C… NA      
#> 4 ReCiPe Midpoint (H) w/o LT - marine eutrophication w/o LT, M… T10: E…  2.62e+5
#> 5 ReCiPe Midpoint (H) w/o LT - marine eutrophication w/o LT, M… T14: P…  1.44e+3
#> 6 ReCiPe Midpoint (H) w/o LT - marine eutrophication w/o LT, M… T15: S…  1.40e-2
#> 7 ReCiPe Midpoint (H) w/o LT - marine eutrophication w/o LT, M… T21: f…  1.06e+4
#> 8 ReCiPe Midpoint (H) w/o LT - marine eutrophication w/o LT, M… T31: E… -1.71e+2
#> 9 ReCiPe Midpoint (H) w/o LT - marine eutrophication w/o LT, M… T4: El…  1.39e+3
#> # … with abbreviated variable name ¹​`0_Reference_Agri`
#>