Make pivot data

Make pivot data

pivot_data(rawdata_grouped, cols_to_ignore = "unit", key_col = "model",
  value_col = "quantity_sum")

Arguments

rawdata_grouped
grouped rawdata as retrieved by function group_data()

cols_to_ignore
column names not to be considered for transforming the data from the "list" to the "wide" format with tidyr::spread() (default: "unit")

key_col
column name containing the key for transforming the data from the "list" to the "wide" format with tidyr::spread() (default: "model")

value_col
column name containing the values that will be used during transforming the data from the "list" to the "wide" format with tidyr::spread() (default: "quantity_sum")

Value

data.frame with the aggregated results for all different"lci_method"

Examples

zipfile <- system.file("extdata/umberto-nxt_v7.1.0.13.503/Beispiel_Auswertung.zip",
package = "kwb.umberto")
umberto7_csv_dir <- file.path(tempdir(), "Beispiel_Auswertung")
unzip(zipfile, exdir = umberto7_csv_dir)
umberto7_rawdata <- kwb.umberto::import_rawdata(csv_dir = umberto7_csv_dir)

#> Importing csv file '/tmp/Rtmpe2qxXq/Beispiel_Auswertung/Ref_High_C+high N.csv'

#> Using ',' as decimal and '.' as grouping mark. Use read_delim() for more control.

#> Parsed with column specification:
#> cols(
#>   Project = col_character(),
#>   Model = col_character(),
#>   Net = col_character(),
#>   Timestamp = col_character(),
#>   Product = col_character(),
#>   `LCI Method` = col_character(),
#>   Phase = col_character(),
#>   Process = col_character(),
#>   `Material Type` = col_character(),
#>   Material = col_character(),
#>   Quantity = col_double(),
#>   Unit = col_character()
#> )

#> Importing csv file '/tmp/Rtmpe2qxXq/Beispiel_Auswertung/Ref_High_C+N.csv'

#> Using ',' as decimal and '.' as grouping mark. Use read_delim() for more control.

#> Parsed with column specification:
#> cols(
#>   Project = col_character(),
#>   Model = col_character(),
#>   Net = col_character(),
#>   Timestamp = col_character(),
#>   Product = col_character(),
#>   `LCI Method` = col_character(),
#>   Phase = col_character(),
#>   Process = col_character(),
#>   `Material Type` = col_character(),
#>   Material = col_character(),
#>   Quantity = col_double(),
#>   Unit = col_character()
#> )

#> Importing csv file '/tmp/Rtmpe2qxXq/Beispiel_Auswertung/Ref_Low_C+high N.csv'

#> Using ',' as decimal and '.' as grouping mark. Use read_delim() for more control.

#> Parsed with column specification:
#> cols(
#>   Project = col_character(),
#>   Model = col_character(),
#>   Net = col_character(),
#>   Timestamp = col_character(),
#>   Product = col_character(),
#>   `LCI Method` = col_character(),
#>   Phase = col_character(),
#>   Process = col_character(),
#>   `Material Type` = col_character(),
#>   Material = col_character(),
#>   Quantity = col_double(),
#>   Unit = col_character()
#> )

#> Importing csv file '/tmp/Rtmpe2qxXq/Beispiel_Auswertung/Ref_Low_C+N.csv'

#> Using ',' as decimal and '.' as grouping mark. Use read_delim() for more control.

#> Parsed with column specification:
#> cols(
#>   Project = col_character(),
#>   Model = col_character(),
#>   Net = col_character(),
#>   Timestamp = col_character(),
#>   Product = col_character(),
#>   `LCI Method` = col_character(),
#>   Phase = col_character(),
#>   Process = col_character(),
#>   `Material Type` = col_character(),
#>   Material = col_character(),
#>   Quantity = col_double(),
#>   Unit = col_character()
#> )

umberto7_data_grouped <- kwb.umberto::group_data(umberto7_rawdata)
umberto7_data_pivot <- kwb.umberto::pivot_data(umberto7_data_grouped)
head(umberto7_data_pivot)

#> # A tibble: 6 x 6
#> # Groups:   lci_method, process [6]
#>   lci_method process `0a_Ref_Low_C+N` `0b_Ref_High_C+… `0c_Ref_Low_C+h…
#>   <chr>      <chr>              <dbl>            <dbl>            <dbl>
#> 1 cumulativ… T11: E…          145350.          135761.          162218.
#> 2 cumulativ… T12: E…           30204.           29674.           30950.
#> 3 cumulativ… T13: F…           17734.           23087.           25930.
#> 4 cumulativ… T14: P…           21443.           22178.           21871.
#> 5 cumulativ… T15: S…           22137.           22794.           23126.
#> 6 cumulativ… T21: M…            5687.            5829.            5906.
#> # … with 1 more variable: `0d_Ref_High_C+high N` <dbl>

umberto10_csv_dir <- system.file("extdata/umberto-lca+_v10.1.0.3.146",
package = "kwb.umberto")
umberto10_rawdata <- kwb.umberto::import_rawdata(csv_dir = umberto10_csv_dir)

#> Importing csv file '/home/travis/build/KWB-R/kwb.umberto/inst/extdata/umberto-lca+_v10.1.0.3.146/smartech2_model-v0.1.0_input-v0.3.1.csv'

#> Using ',' as decimal and '.' as grouping mark. Use read_delim() for more control.

#> Parsed with column specification:
#> cols(
#>   Project = col_character(),
#>   Model = col_character(),
#>   Net = col_character(),
#>   Timestamp = col_character(),
#>   Product = col_character(),
#>   `Product Name` = col_character(),
#>   `Product Arrow` = col_character(),
#>   `Product Flow Amount` = col_character(),
#>   `LCI Method` = col_character(),
#>   Phase = col_character(),
#>   Process = col_character(),
#>   `Material Type` = col_character(),
#>   Material = col_character(),
#>   Quantity = col_double(),
#>   Unit = col_character()
#> )

umberto10_data_grouped <- kwb.umberto::group_data(umberto10_rawdata)
umberto10_data_pivot <- kwb.umberto::pivot_data(umberto10_data_grouped)
head(umberto10_data_pivot)

#> # A tibble: 6 x 3
#> # Groups:   lci_method, process [6]
#>   lci_method                                 process            `0_Reference_Ag…
#>   <chr>                                      <chr>                         <dbl>
#> 1 cumulative energy demand - fossil, non-re… T14: Polymer              10433117.
#> 2 cumulative energy demand - fossil, non-re… T15: Sludge trans…            1763.
#> 3 cumulative energy demand - fossil, non-re… T21: fertilizer s…        -7072323.
#> 4 cumulative energy demand - fossil, non-re… T31: Electricity …       -13154124.
#> 5 cumulative energy demand - fossil, non-re… T4: Electricity […       181431756.
#> 6 cumulative energy demand - nuclear, non-r… T14: Polymer                296075.