Make pivot data

Make pivot data

Usage

pivot_data(
  rawdata_grouped,
  cols_to_ignore = "unit",
  key_col = "model",
  value_col = "quantity_sum"
)

Arguments

rawdata_grouped

grouped rawdata as retrieved by function group_data()

cols_to_ignore

column names not to be considered for transforming the data from the "list" to the "wide" format with tidyr::spread() (default: "unit")

key_col

column name containing the key for transforming the data from the "list" to the "wide" format with tidyr::spread() (default: "model")

value_col

column name containing the values that will be used during transforming the data from the "list" to the "wide" format with tidyr::spread() (default: "quantity_sum")

Value

data.frame with the aggregated results for all different"lci_method"

Examples

zipfile <- system.file("extdata/umberto-nxt_v7.1.0.13.503/Beispiel_Auswertung.zip", 
package = "kwb.umberto")
umberto7_csv_dir <- file.path(tempdir(), "Beispiel_Auswertung")
unzip(zipfile, exdir = umberto7_csv_dir)
umberto7_rawdata <- kwb.umberto::import_rawdata(csv_dir = umberto7_csv_dir)
#> Importing csv file 'C:\Users\RUNNER~1\AppData\Local\Temp\RtmpWCmvdA/Beispiel_Auswertung/Ref_High_C+N.csv'
#> ℹ Using "','" as decimal and "'.'" as grouping mark. Use `read_delim()` for more control.
#> Rows: 19996 Columns: 12
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ";"
#> chr (11): Project, Model, Net, Timestamp, Product, LCI Method, Phase, Proces...
#> dbl  (1): Quantity
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Importing csv file 'C:\Users\RUNNER~1\AppData\Local\Temp\RtmpWCmvdA/Beispiel_Auswertung/Ref_High_C+high N.csv'
#> ℹ Using "','" as decimal and "'.'" as grouping mark. Use `read_delim()` for more control.
#> Rows: 19996 Columns: 12
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ";"
#> chr (11): Project, Model, Net, Timestamp, Product, LCI Method, Phase, Proces...
#> dbl  (1): Quantity
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Importing csv file 'C:\Users\RUNNER~1\AppData\Local\Temp\RtmpWCmvdA/Beispiel_Auswertung/Ref_Low_C+N.csv'
#> ℹ Using "','" as decimal and "'.'" as grouping mark. Use `read_delim()` for more control.
#> Rows: 19996 Columns: 12
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ";"
#> chr (11): Project, Model, Net, Timestamp, Product, LCI Method, Phase, Proces...
#> dbl  (1): Quantity
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Importing csv file 'C:\Users\RUNNER~1\AppData\Local\Temp\RtmpWCmvdA/Beispiel_Auswertung/Ref_Low_C+high N.csv'
#> ℹ Using "','" as decimal and "'.'" as grouping mark. Use `read_delim()` for more control.
#> Rows: 19996 Columns: 12
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ";"
#> chr (11): Project, Model, Net, Timestamp, Product, LCI Method, Phase, Proces...
#> dbl  (1): Quantity
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
umberto7_data_grouped <- kwb.umberto::group_data(umberto7_rawdata)
umberto7_data_pivot <- kwb.umberto::pivot_data(umberto7_data_grouped)
head(umberto7_data_pivot)
#> # A tibble: 6 × 6
#> # Groups:   lci_method, process [6]
#>   lci_method                             process 0a_Re…¹ 0b_Re…² 0c_Re…³ 0d_Re…⁴
#>   <chr>                                  <chr>     <dbl>   <dbl>   <dbl>   <dbl>
#> 1 ReCiPe Midpoint (H) w/o LT - climate … T03: A…  25498.   6656.  17548.   2705.
#> 2 ReCiPe Midpoint (H) w/o LT - climate … T06: C…    607.    604.    615.    608.
#> 3 ReCiPe Midpoint (H) w/o LT - climate … T07: C…    277.    281.    284.    282.
#> 4 ReCiPe Midpoint (H) w/o LT - climate … T11: E…  12520.  11694.  13973.  13312.
#> 5 ReCiPe Midpoint (H) w/o LT - climate … T12: E…   2602.   2556.   2666.   2569.
#> 6 ReCiPe Midpoint (H) w/o LT - climate … T13: F…   1548.   2016.   2264.   2085.
#> # … with abbreviated variable names ¹​`0a_Ref_Low_C+N`, ²​`0b_Ref_High_C+N`,
#> #   ³​`0c_Ref_Low_C+high N`, ⁴​`0d_Ref_High_C+high N`

umberto10_csv_dir <- system.file("extdata/umberto-lca+_v10.1.0.3.146", 
package = "kwb.umberto")
umberto10_rawdata <- kwb.umberto::import_rawdata(csv_dir = umberto10_csv_dir)
#> Importing csv file 'D:/a/_temp/Library/kwb.umberto/extdata/umberto-lca+_v10.1.0.3.146/smartech2_model-v0.1.0_input-v0.3.1.csv'
#> ℹ Using "','" as decimal and "'.'" as grouping mark. Use `read_delim()` for more control.
#> Rows: 8456 Columns: 15
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ";"
#> chr (14): Project, Model, Net, Timestamp, Product, Product Name, Product Arr...
#> dbl  (1): Quantity
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
umberto10_data_grouped <- kwb.umberto::group_data(umberto10_rawdata)
umberto10_data_pivot <- kwb.umberto::pivot_data(umberto10_data_grouped)
head(umberto10_data_pivot)
#> # A tibble: 6 × 3
#> # Groups:   lci_method, process [6]
#>   lci_method                                                     process 0_Ref…¹
#>   <chr>                                                          <chr>     <dbl>
#> 1 ReCiPe Midpoint (H) w/o LT - climate change w/o LT, GWP100 w/… T03: A…  3.78e6
#> 2 ReCiPe Midpoint (H) w/o LT - climate change w/o LT, GWP100 w/… T06: C…  1.81e5
#> 3 ReCiPe Midpoint (H) w/o LT - climate change w/o LT, GWP100 w/… T07: C…  2.89e5
#> 4 ReCiPe Midpoint (H) w/o LT - climate change w/o LT, GWP100 w/… T14: P…  3.93e5
#> 5 ReCiPe Midpoint (H) w/o LT - climate change w/o LT, GWP100 w/… T15: S…  1.10e2
#> 6 ReCiPe Midpoint (H) w/o LT - climate change w/o LT, GWP100 w/… T21: f… -2.17e5
#> # … with abbreviated variable name ¹​`0_Reference_Agri`